+++ src/drf_yasg/openapi.py
@@ -35,6 +35,7 @@ FORMAT_URI = "uri"  #:
 # pulled out of my ass
 FORMAT_UUID = "uuid"  #:
 FORMAT_SLUG = "slug"  #:
+FORMAT_DECIMAL = "decimal"
 
 IN_BODY = 'body'  #:
 IN_PATH = 'path'  #: